[rename] global 패키지 내부에 존재하던 파일들 단순 경로 수정(CUL-173)
# 수정의의
# Domain모듈을 보면 core와 같은 레벨에 common 패키지가 존재하는데, 일관성을 맞춰주고자 패키지 경로 수정
# 이로인해 core패키지는 도메인만 남아 책임과 역할 부분에 있어서 더 명확해진다고 생각합니다.
# etc. Domain 모듈의 경우에도 core하위에 global이 존재하는데 이것도 차후에 옮기면 어떨까 고민중입니다.
