+++ CultureSpot-ServiceApi/src/main/java/com/culturespot/culturespotserviceapi/common/utils/response/ApiResponse.java
@@ -1,0 +1,5 @@
+package com.culturespot.culturespotserviceapi.common.utils.response;
+
+public class ApiResponse {
+}
+
