+++ CultureSpot-ServiceApi/src/main/java/com/culturespot/culturespotserviceapi/core/auth/controller/TestController.java
@@ -3,7 +3,7 @@ package com.culturespot.culturespotserviceapi.core.auth.controller;
 import com.culturespot.culturespotdomain.core.user.entity.User;
 import com.culturespot.culturespotserviceapi.core.auth.annotation.Auth;
 import com.culturespot.culturespotserviceapi.core.auth.annotation.MemberOnly;
-import com.culturespot.culturespotserviceapi.core.global.security.endpoint.EndpointType;
+import com.culturespot.culturespotserviceapi.common.security.endpoint.EndpointType;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.GetMapping;
