--- CultureSpot-ServiceApi/src/main/java/com/culturespot/culturespotserviceapi/core/global/utils/response/ApiResponse.java
@@ -1,5 +1,0 @@
-package com.culturespot.culturespotserviceapi.core.global.utils.response;
-
-public class ApiResponse {
-}
-
